refactor: split our plugin to two separated plugins
